backend-ai-agent/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and LLM integration
ollama==0.1.7
langchain==0.0.340
langchain-community==0.0.1

# Code analysis and parsing
ast==0.0.2
libcst==1.1.0
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.1
tree-sitter-java==0.20.2

# Vector database
chromadb==0.4.18

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2